refactor(script): Phase 4 Step 1 - Extract git_operations.py
- Create new module src/script/git_operations.py (291 lines)
- Extract 3 functions from script_manager.py:
  - execute_git_script() - CI-safe wrapper
  - clone_git_repo() - Repository cloning
  - execute_git_script_new_method() - NEW METHOD (2024+ automation)
- Update script_manager.py to import from git_operations
- Reduce script_manager.py: 1,884 → 1,646 lines (-238 lines, 12.6%)
- Maintain 100% backward compatibility
- All imports working correctly
