+++ src/script/script_manager.py
@@ -11,6 +11,13 @@ from src.utils.app_logger import logger
 from src.utils.git_script_automator import GitScriptAutomator, EdgeAutomator, ChromeAutomator
 from typing import Dict as _DictReturn
 
+# Import git operations from new module
+from src.script.git_operations import (
+    execute_git_script,
+    clone_git_repo,
+    execute_git_script_new_method,
+)
+
 # Initialize directory migration on module import
 try:
     from src.runner.migration_tool import migrate_user_scripts
@@ -27,72 +34,6 @@ except ImportError:
 except Exception as e:
     logger.error(f"❌ Migration initialization failed: {e}")
 
-
-async def execute_git_script(
-    url: str,
-    goal: str,
-    browser_type: Optional[str] = None,
-    action_type: str = "git-script",
-    headless: bool = True,
-    save_recording_path: Optional[str] = None,
-) -> _DictReturn[str, Any]:
-    """CI-safe wrapper for executing a git-script.
-
-    This function exists to satisfy tests that import `execute_git_script` from
-    src.script.script_manager. For CI safety and to avoid launching real
-    browsers during unit tests, this implementation returns a stubbed result.
-
-    Returns a dict with at least keys: success (bool), status (str), error (str).
-    """
-    # Provide a deterministic, non-executing response suitable for CI tests.
-    return {
-        "success": False,
-        "status": "skipped",
-        "error": "CI-safe stub: execute_git_script is not executed in unit tests",
-        "url": url,
-        "goal": goal,
-        "browser_type": browser_type or "chrome",
-        "action_type": action_type,
-        "headless": headless,
-    }
-
-async def clone_git_repo(git_url: str, version: str = 'main', target_dir: Optional[str] = None) -> str:
-    """Clone a git repository and checkout specified version/branch"""
-    if target_dir is None:
-        target_dir = tempfile.mkdtemp()
-    
-    try:
-        # If the directory already exists, handle it appropriately
-        if os.path.exists(target_dir):
-            if os.path.exists(os.path.join(target_dir, '.git')):
-                # It's a git repo, try to pull the latest changes
-                logger.info(f"Repository already exists at {target_dir}, pulling latest changes")
-                subprocess.run(['git', 'fetch', '--all'], cwd=target_dir, check=True)
-                subprocess.run(['git', 'reset', '--hard', 'origin/main'], cwd=target_dir, check=True)
-            else:
-                # Directory exists but is not a git repo, remove it and clone
-                logger.info(f"Directory exists but is not a git repo, removing: {target_dir}")
-                shutil.rmtree(target_dir)
-                os.makedirs(target_dir, exist_ok=True)
-                logger.info(f"Cloning repository from {git_url} to {target_dir}")
-                subprocess.run(['git', 'clone', git_url, target_dir], check=True)
-        else:
-            # Directory doesn't exist, create it and clone
-            os.makedirs(os.path.dirname(target_dir), exist_ok=True)
-            logger.info(f"Cloning repository from {git_url} to {target_dir}")
-            subprocess.run(['git', 'clone', git_url, target_dir], check=True)
-        
-        # Checkout specific version/branch if provided
-        if version:
-            logger.info(f"Checking out version/branch: {version}")
-            subprocess.run(['git', 'checkout', version], cwd=target_dir, check=True)
-        
-        return target_dir
-    except subprocess.SubprocessError as e:
-        logger.error(f"Git operation failed: {str(e)}")
-        if os.path.exists(target_dir):
-            shutil.rmtree(target_dir)
-        raise RuntimeError(f"Failed to clone repository: {str(e)}")
 
 def generate_browser_script(script_info: Dict[str, Any], params: Dict[str, str], headless: bool = False) -> str:
     """
@@ -1703,182 +1644,3 @@ async def move_script_files_to_artifacts(script_info: Dict[str, Any], script_pat
         logger.error(f"❌ Failed to archive script files to artifacts: {e}")
         # Don't raise exception to avoid breaking the main execution flow
 
-async def execute_git_script_new_method(
-    script_info: Dict[str, Any], 
-    params: Dict[str, str], 
-    headless: bool, 
-    save_recording_path: Optional[str],
-    browser_type: Optional[str]
-) -> Tuple[str, Optional[str]]:
-    """
-    Execute git-script using the NEW METHOD (2024+ stable automation)
-    """
-    logger.info("🎯 Executing git-script with NEW METHOD (2024+ ProfileManager + BrowserLauncher)")
-    
-    git_url = script_info['git']
-    script_path = script_info['script_path']
-    version = script_info.get('version', 'main')
-    
-    try:
-        # If running under pytest/CI or explicit test mode, use the CI-safe stub
-        test_mode = os.getenv('PYTEST_CURRENT_TEST') is not None or os.getenv('CI', '').lower() == 'true' or os.getenv('BYKILT_TEST_MODE') == '1'
-        if test_mode:
-            logger.info("ℹ️ Detected test mode/CI - using CI-safe git-script stub")
-            # Use the lightweight CI-safe stub defined earlier in this module
-            stub = await execute_git_script(git_url, script_info.get('name', 'git-script'), browser_type=browser_type, headless=headless, save_recording_path=save_recording_path)
-            # Map stub result into the (message, path) contract expected by callers
-            status = stub.get('status', 'skipped')
-            msg = f"git-script stubbed: {status}"
-            return msg, None
-
-        # Step 1: Clone git repository
-        repo_dir = os.path.join(tempfile.gettempdir(), 'bykilt_gitscripts', Path(git_url).stem)
-        os.makedirs(os.path.dirname(repo_dir), exist_ok=True)
-        repo_dir = await clone_git_repo(git_url, version, repo_dir)
-        
-        # Step 2: Validate and normalize script path (GIT_SCRIPT_V2 feature flag)
-        git_script_v2_enabled = os.getenv('GIT_SCRIPT_V2', 'false').lower() == 'true'
-        
-        if git_script_v2_enabled:
-            from src.utils.git_script_path import validate_git_script_path
-            try:
-                full_script_path, validation_result = validate_git_script_path(repo_dir, script_path)
-                logger.info(f"✅ Script path validated (V2): {script_path} -> {full_script_path}")
-                if validation_result.security_warnings:
-                    logger.warning(f"⚠️ Security warnings: {validation_result.security_warnings}")
-            except Exception as e:
-                raise ValueError(f"Script path validation failed: {e}")
-        else:
-            # Legacy path handling
-            full_script_path = os.path.join(repo_dir, script_path)
-            logger.info(f"ℹ️ Using legacy path handling: {script_path} -> {full_script_path}")
-        
-        if not os.path.exists(full_script_path):
-            raise FileNotFoundError(f"Script not found at path: {full_script_path}")
-        
-        # Step 2: Determine browser type
-        from src.browser.browser_config import browser_config
-        override_browser = os.environ.get('BYKILT_OVERRIDE_BROWSER_TYPE')
-        if browser_type:
-            current_browser = browser_type
-        elif override_browser:
-            current_browser = override_browser
-        else:
-            current_browser = browser_config.get_current_browser()
-        
-        logger.info(f"🎯 NEW METHOD using browser: {current_browser}")
-        
-        # Step 3: Initialize NEW METHOD automator
-        if current_browser.lower() == 'edge':
-            automator = EdgeAutomator()
-        elif current_browser.lower() == 'chrome':
-            automator = ChromeAutomator()
-        else:
-            # Fallback to Edge for unknown types
-            logger.warning(f"⚠️ Unknown browser type {current_browser}, falling back to Edge")
-            automator = EdgeAutomator()
-        
-        # Step 4: Validate source profile
-        if not automator.validate_source_profile():
-            raise ValueError(f"Source profile validation failed for {current_browser}")
-        
-        # Step 5: Execute workflow
-        workspace_dir = os.path.dirname(full_script_path)
-        
-        # Initialize recording if enabled
-        recording_context = None
-        if save_recording_path or True:  # Enable recording by default for git-script
-            from src.utils.recording_factory import RecordingFactory
-            
-            # Get the main workspace directory
-            main_workspace = Path(__file__).parent.parent.resolve()
-            
-            # Use explicit absolute path for recording
-            actual_save_recording_path = save_recording_path
-            if not actual_save_recording_path:
-                actual_save_recording_path = str(main_workspace / 'artifacts' / 'runs' / f"{Path(git_url).stem}-{version}-art" / 'videos')
-            
-            run_context = {
-                'run_id': f"{Path(git_url).stem}-{version}",
-                'run_type': 'git-script',
-                'save_recording_path': actual_save_recording_path,
-                'enable_recording': True
-            }
-            recording_context = RecordingFactory.init_recorder(run_context)
-            logger.info(f"🎥 Recording initialized for git-script: {actual_save_recording_path}")
-        
-        # For NEW METHOD, we force headful mode for Edgebrowser stability
-        # 2024+ finding: Edge headless mode is fundamentally unstable
-        if current_browser.lower() == 'edge':
-            actual_headless = False  # Force headful for Edge
-            if headless:
-                logger.warning("⚠️ Edge headless forced to headful for stability (NEW METHOD)")
-        else:
-            actual_headless = headless
-        
-        # Execute complete automation workflow
-        result = await automator.execute_git_script_workflow(
-            workspace_dir=workspace_dir,
-            script_path=full_script_path,
-            command=script_info.get('command', f'python {script_path}'),
-            params=params
-        )
-        
-        # Set recording path in environment for the executed script
-        if recording_context:
-            # For git-script, create a local recording directory within the cloned repository
-            # to ensure the script can write recordings regardless of working directory
-            try:
-                workspace_path = Path(workspace_dir)
-                local_recording_dir = workspace_path / 'tmp' / 'record_videos'
-                local_recording_dir.mkdir(parents=True, exist_ok=True)
-                
-                # Set the environment variable for the script execution
-                os.environ['RECORDING_PATH'] = str(local_recording_dir)
-                logger.info(f"🎥 Recording path set for git-script execution: {local_recording_dir} (local to workspace)")
-            except Exception as e:
-                # Fallback to original absolute path approach
-                os.environ['RECORDING_PATH'] = str(recording_context.recording_path)
-                logger.warning(f"⚠️ Failed to create local recording directory, using absolute path: {recording_context.recording_path} (error: {e})")
-        
-        if result["success"]:
-            # After successful execution, copy recording files to the main artifacts directory
-            if recording_context:
-                try:
-                    workspace_path = Path(workspace_dir)
-                    local_recording_dir = workspace_path / 'tmp' / 'record_videos'
-                    target_recording_dir = Path(recording_context.recording_path)
-                    
-                    if local_recording_dir.exists() and target_recording_dir.exists():
-                        import shutil
-                        # Copy all recording files from local directory to target directory
-                        for recording_file in local_recording_dir.glob("*.webm"):
-                            target_file = target_recording_dir / recording_file.name
-                            shutil.copy2(recording_file, target_file)
-                            logger.info(f"📹 Copied recording file: {recording_file.name} -> {target_file}")
-                        
-                        # Also copy mp4 files if they exist
-                        for recording_file in local_recording_dir.glob("*.mp4"):
-                            target_file = target_recording_dir / recording_file.name
-                            shutil.copy2(recording_file, target_file)
-                            logger.info(f"📹 Copied recording file: {recording_file.name} -> {target_file}")
-                        
-                        logger.info(f"✅ Recording files copied from {local_recording_dir} to {target_recording_dir}")
-                    else:
-                        logger.warning(f"⚠️ Recording directories not found: local={local_recording_dir.exists()}, target={target_recording_dir.exists()}")
-                except Exception as e:
-                    logger.warning(f"⚠️ Failed to copy recording files: {e}")
-            
-            success_msg = f"NEW METHOD git-script executed successfully with {current_browser}"
-            logger.info(f"✅ {success_msg}")
-            logger.info(f"📁 SeleniumProfile: {result.get('selenium_profile')}")
-            return success_msg, full_script_path
-        else:
-            error_msg = f"NEW METHOD git-script failed: {result.get('error', 'Unknown error')}"
-            logger.error(f"❌ {error_msg}")
-            return error_msg, None
-            
-    except Exception as e:
-        error_msg = f"NEW METHOD git-script execution failed: {str(e)}"
-        logger.error(f"❌ {error_msg}")
-        return error_msg, None
